requirements.txt
PyMuPDF
pytesseract
beautifulsoup4
Pillow
